+++ src/main/java/chess/domain/game/ChessGame.java
@@ -2,12 +2,10 @@ package chess.domain.game;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Position;
-import chess.domain.piece.Color;
+import chess.domain.piece.feature.Color;
 import chess.domain.piece.Piece;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class ChessGame {
 	public static final String TURN_MESSAGE = "%s의 차례입니다.";
