+++ src/main/java/chess/domain/piece/Blank.java
@@ -3,6 +3,7 @@ package chess.domain.piece;
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
+import chess.domain.piece.feature.Type;
 
 import java.util.List;
