+++ src/main/java/chess/domain/piece/Piece.java
@@ -3,6 +3,8 @@ package chess.domain.piece;
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
+import chess.domain.piece.feature.Color;
+import chess.domain.piece.feature.Type;
 
 import java.util.List;
 import java.util.Objects;
