+++ src/main/java/chess/domain/piece/fixeddistance/FixedDistancePiece.java
@@ -1,8 +1,10 @@
-package chess.domain.piece;
+package chess.domain.piece.fixeddistance;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
+import chess.domain.piece.Piece;
+import chess.domain.piece.feature.Color;
 
 public abstract class FixedDistancePiece extends Piece {
 	public FixedDistancePiece(final Color color, final Position position) {
