+++ src/main/java/chess/domain/piece/flexibledistance/FlexibleDistancePiece.java
@@ -1,8 +1,10 @@
-package chess.domain.piece;
+package chess.domain.piece.flexibledistance;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
+import chess.domain.piece.Piece;
+import chess.domain.piece.feature.Color;
 
 public abstract class FlexibleDistancePiece extends Piece {
 	public FlexibleDistancePiece(final Color color, final Position position) {
