+++ src/test/java/chess/domain/game/ChessGameTest.java
@@ -3,7 +3,7 @@ package chess.domain.game;
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Position;
 import chess.domain.piece.Blank;
-import chess.domain.piece.Color;
+import chess.domain.piece.feature.Color;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
