+++ src/test/java/chess/domain/game/ResultTest.java
@@ -1,12 +1,7 @@
 package chess.domain.game;
 
-import chess.domain.piece.Color;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
