+++ src/test/java/chess/domain/piece/BlankTest.java
@@ -3,6 +3,7 @@ package chess.domain.piece;
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Direction;
 import chess.domain.board.Position;
+import chess.domain.piece.feature.Color;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
