+++ src/test/java/chess/domain/piece/KingTest.java
@@ -2,6 +2,8 @@ package chess.domain.piece;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Position;
+import chess.domain.piece.feature.Color;
+import chess.domain.piece.fixeddistance.King;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
