+++ src/test/java/chess/domain/piece/RookTest.java
@@ -2,6 +2,9 @@ package chess.domain.piece;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.Position;
+import chess.domain.piece.feature.Color;
+import chess.domain.piece.fixeddistance.Knight;
+import chess.domain.piece.flexibledistance.Rook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
